SAK-40067: The averaged peer review grade is rounded down

+++ assignment/impl/src/java/org/sakaiproject/assignment/impl/AssignmentPeerAssessmentServiceImpl.java
@@ -447,7 +447,7 @@ public class AssignmentPeerAssessmentServiceImpl extends HibernateDaoSupport imp
                         }
                     }
                     if (denominator > 0) {
-                        totalScore = Math.round(totalScore / denominator);
+                        totalScore = Math.round(totalScore.floatValue() / denominator);
                     } else {
                         totalScore = null;
                     }
